requirements/base.txt
Django==1.8.1
Jinja2==2.8
MarkupSafe==0.23
Pillow==2.8.1
Pygments==2.0.2
astroid==1.4.2
colorama==0.3.5
dj-database-url==0.3.0
dj-static==0.0.6
dj-stripe==0.8.0
django-angular==0.7.13
django-appconf==1.0.1
django-bootstrap3==5.4.0
django-braces==1.8.1
django-debug-toolbar==1.3.0
django-extensions==1.5.3
django-grappelli==2.6.4
django-guardian==1.3
django-hstore==1.4
django-lint==2.0.1
django-mailgun==0.9.1
django-markupfield==1.3.5
django-model-utils==2.4
django-sortedm2m==1.0.2
django-tastypie==0.12.2
django-tinymce==2.0.4
django-userena==1.4.0
djrill==1.3.0
docopt==0.4.0
easy-thumbnails==2.2
ecdsa==0.13
gunicorn==19.3.0
hanzidentifier==1.0.2
html2text==2015.4.14
ipython==3.2.1
jieba==0.37
jsonfield==1.0.3
jsonschema==2.5.1
lazy-object-proxy==1.2.1
mandrill==1.0.57
mistune==0.7
paramiko==1.15.2
psycopg2==2.6
ptyprocess==0.5
pycrypto==2.6.1
pylint==1.5.2
python-dateutil==2.4.2
python-mimeparse==0.1.4
pytz==2015.7
pyzmq==14.7.0
raven==6.0.0
requests==2.7.0
six==1.9.0
sqlparse==0.1.15
static3==0.5.1
stripe==1.29.1
terminado==0.5
tornado==4.2.1
whitenoise==1.0.6
wrapt==1.10.6
xpinyin==0.5.3
zhon==1.1.4
xlrd==1.0.0
git+https://github.com/pjdelport/django-seo-js.git@update-for-newer-Python-and-Django